Extend use of function to navigate to move
